refactor: move inner wrapper to new component
