+++ src/data-workspace/data-entry-cell/data-entry-cell.js
@@ -1,5 +1,3 @@
-import { IconMore16, colors } from '@dhis2/ui'
-import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 import { useField } from 'react-final-form'
@@ -10,36 +8,10 @@ import {
     getDataSetById,
 } from '../../metadata/selectors.js'
 import styles from './data-entry-cell.module.css'
+import { InnerWrapper } from './inner-wrapper.js'
 import { useDataValueMutation } from './use-data-value-mutation.js'
 import { ValidationTooltip } from './validation-tooltip.js'
 import { getInputByDataElement, VALUE_TYPES } from './value-types.js'
-
-/** Three dots or triangle in top-right corner of cell */
-const SyncStatusIndicator = ({ isLoading, isSynced }) => {
-    return (
-        <div className={styles.topRightIndicator}>
-            {isLoading ? (
-                <IconMore16 color={colors.grey700} />
-            ) : isSynced ? (
-                <div className={styles.topRightTriangle} />
-            ) : null}
-        </div>
-    )
-}
-SyncStatusIndicator.propTypes = {
-    isLoading: PropTypes.bool,
-    isSynced: PropTypes.bool,
-}
-
-/** Grey triangle in bottom left of cell */
-const CommentIndicator = ({ isComment }) => {
-    return (
-        <div className={styles.bottomLeftIndicator}>
-            {isComment && <div className={styles.bottomLeftTriangle} />}
-        </div>
-    )
-}
-CommentIndicator.propTypes = { isComment: PropTypes.bool }
 
 export const DataEntryCell = React.memo(function DataEntryCell({
     dataElement: de,
@@ -117,43 +89,24 @@ export const DataEntryCell = React.memo(function DataEntryCell({
         )
     }
 
-    const onKeyDown = (event) => {
-        // field navigation is handled by KeyboardNavManager
-        const { key, shiftKey } = event
-        if (key === 'Enter' && shiftKey) {
-            // todo: open data item details
-        }
-    }
-
     // todo: get data details (via getDataValue?)
     // todo: implement read-only cells
 
+    // todo: refactor this into lower components
     const isFileInput =
         de.valueType === 'FILE_RESOURCE' || de.valueType === 'IMAGE'
-
     const isSynced = isFileInput
         ? !isFileLoading && isFileSynced
         : meta.valid && !isIdle && !isLoading && !isError
-    const cellStateClassName = meta.invalid
-        ? styles.invalid
-        : isSynced
-        ? styles.synced
-        : null
 
     const Input = getInputByDataElement(de)
 
     return (
         <td className={styles.dataEntryCell}>
             <ValidationTooltip fieldname={fieldname}>
-                <div
-                    className={cx(styles.cellInnerWrapper, cellStateClassName, {
-                        [styles.active]: meta.active,
-                        [styles.disabled]: false, // todo
-                    })}
-                >
+                <InnerWrapper fieldname={fieldname} isSynced={isSynced}>
                     <Input
                         name={fieldname}
-                        onKeyDown={onKeyDown}
                         // disabled={true} (todo)
                         // props for most inputs:
                         syncData={syncData}
@@ -164,13 +117,7 @@ export const DataEntryCell = React.memo(function DataEntryCell({
                         setIsFileSynced={setIsFileSynced}
                         setIsFileLoading={setIsFileLoading}
                     />
-                    <SyncStatusIndicator
-                        isLoading={isLoading || isFileLoading}
-                        isSynced={isSynced}
-                    />
-                    {/* todo: show indicator if there is a comment */}
-                    <CommentIndicator isComment={false} />
-                </div>
+                </InnerWrapper>
             </ValidationTooltip>
         </td>
     )
